+++ processor/src/test/resources/expected/com/example/prop/React4j_ObservableViaMemoizeProp.java
@@ -1,56 +1,41 @@
 package com.example.prop;
 
-import arez.Arez;
-import arez.Disposable;
 import arez.ObservableValue;
-import arez.Observer;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.Executor;
 import arez.annotations.Feature;
 import arez.annotations.InjectMode;
-import arez.annotations.Memoize;
 import arez.annotations.Observable;
 import arez.annotations.ObservableValueRef;
-import arez.annotations.Observe;
-import arez.annotations.ObserverRef;
-import arez.annotations.Priority;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsConstructor;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
-import org.realityforge.braincheck.Guards;
 import react4j.React;
 import react4j.ReactNode;
 import react4j.internal.ComponentConstructorFunction;
 import react4j.internal.NativeComponent;
-import react4j.internal.OnComponentDidMount;
-import react4j.internal.OnComponentDidUpdate;
 import react4j.internal.OnComponentWillUnmount;
 import react4j.internal.OnShouldComponentUpdate;
-import react4j.internal.arez.ComponentState;
-import react4j.internal.arez.IntrospectUtil;
-import react4j.internal.arez.SchedulerUtil;
 
 @ArezComponent(
     name = "ObservableViaMemoizeProp",
     disposeTrackable = Feature.DISABLE,
+    allowEmpty = true,
     inject = InjectMode.NONE,
     dagger = Feature.DISABLE
 )
 @Generated("react4j.processor.ReactProcessor")
 abstract class React4j_ObservableViaMemoizeProp extends ObservableViaMemoizeProp {
-  private int $$react4j$$_state;
-
   React4j_ObservableViaMemoizeProp(@Nonnull final NativeComponent nativeComponent) {
     bindComponent( nativeComponent );
   }
 
   @Nonnull
   private static ComponentConstructorFunction getConstructorFunction() {
-    final ComponentConstructorFunction componentConstructor = ( React.shouldStoreDebugDataAsState() || React.shouldValidatePropValues() ) ? NativeReactComponent::new : LiteNativeReactComponent::new;
+    final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
     if ( React.enableComponentNames() ) {
       Js.asPropertyMap( componentConstructor ).set( "displayName", "ObservableViaMemoizeProp" );
     }
@@ -86,70 +71,11 @@ abstract class React4j_ObservableViaMemoizeProp extends ObservableViaMemoizeProp
       getValueObservableValue().reportChanged();
       modified = true;
     }
-    return modified || ComponentState.SCHEDULED == $$react4j$$_state;
-  }
-
-  private void $$react4j$$_componentDidMount() {
-    if ( React.shouldStoreDebugDataAsState() ) {
-      storeDebugDataAsState();
-    }
-  }
-
-  private void $$react4j$$_componentDidUpdate() {
-    if ( React.shouldStoreDebugDataAsState() ) {
-      storeDebugDataAsState();
-    }
+    return modified;
   }
 
   private void $$react4j$$_componentWillUnmount() {
-    $$react4j$$_state = ComponentState.UNMOUNTED;
     ((Arez_React4j_ObservableViaMemoizeProp) this).dispose();
-  }
-
-  @Override
-  @Nullable
-  @Observe(
-      name = "render",
-      priority = Priority.LOW,
-      executor = Executor.EXTERNAL,
-      observeLowerPriorityDependencies = true,
-      reportResult = false
-  )
-  protected ReactNode render() {
-    $$react4j$$_state = ComponentState.IDLE;
-    SchedulerUtil.pauseUntilRenderLoopComplete();
-    assert Disposable.isNotDisposed( this );
-    final ReactNode result = super.render();
-    if ( Arez.shouldCheckInvariants() && Arez.areSpiesEnabled() ) {
-      Guards.invariant( () -> !getRenderObserver().getContext().getSpy().asObserverInfo( getRenderObserver() ).getDependencies().isEmpty(), () -> "ReactArezComponent render completed on '" + this + "' but the component does not have any Arez dependencies. This component should extend react4j.Component instead." );
-    }
-    return result;
-  }
-
-  final void onRenderDepsChange() {
-    if ( ComponentState.IDLE == $$react4j$$_state ) {
-      $$react4j$$_state = ComponentState.SCHEDULED;
-      scheduleRender( false );
-    }
-  }
-
-  @Nonnull
-  @ObserverRef
-  abstract Observer getRenderObserver();
-
-  @Override
-  protected final void populateDebugData(@Nonnull final JsPropertyMap<Object> data) {
-    if ( React.shouldStoreDebugDataAsState() && Arez.areSpiesEnabled() ) {
-      IntrospectUtil.collectDependencyDebugData( getRenderObserver(), data );
-    }
-  }
-
-  @Override
-  @Memoize(
-      priority = Priority.LOWEST
-  )
-  int someValue(final int i) {
-    return super.someValue(i);
   }
 
   static final class Factory {
@@ -160,33 +86,7 @@ abstract class React4j_ObservableViaMemoizeProp extends ObservableViaMemoizeProp
     static final String value = React.shouldMinimizePropKeys() ? "a" : "value";
   }
 
-  private static final class LiteNativeReactComponent extends NativeComponent implements OnShouldComponentUpdate, OnComponentWillUnmount {
-    private React4j_ObservableViaMemoizeProp $$react4j$$_component;
-
-    @JsConstructor
-    LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
-      super( props );
-      $$react4j$$_component = new Arez_React4j_ObservableViaMemoizeProp( this );
-    }
-
-    @Override
-    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextProps) {
-      return $$react4j$$_component.$$react4j$$_shouldComponentUpdate( nextProps );
-    }
-
-    @Override
-    public final void componentWillUnmount() {
-      $$react4j$$_component.$$react4j$$_componentWillUnmount();
-    }
-
-    @Override
-    @Nullable
-    public final ReactNode render() {
-      return $$react4j$$_component.render();
-    }
-  }
-
-  private static final class NativeReactComponent extends NativeComponent implements OnComponentDidMount, OnComponentDidUpdate, OnShouldComponentUpdate, OnComponentWillUnmount {
+  private static final class NativeReactComponent extends NativeComponent implements OnShouldComponentUpdate, OnComponentWillUnmount {
     private React4j_ObservableViaMemoizeProp $$react4j$$_component;
 
     @JsConstructor
@@ -196,18 +96,8 @@ abstract class React4j_ObservableViaMemoizeProp extends ObservableViaMemoizeProp
     }
 
     @Override
-    public final void componentDidMount() {
-      $$react4j$$_component.$$react4j$$_componentDidMount();
-    }
-
-    @Override
     public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextProps) {
       return $$react4j$$_component.$$react4j$$_shouldComponentUpdate( nextProps );
-    }
-
-    @Override
-    public final void componentDidUpdate(@Nonnull final JsPropertyMap<Object> prevProps) {
-      $$react4j$$_component.$$react4j$$_componentDidUpdate();
     }
 
     @Override